Migrated geometry function sets as core geometry.
