+++ ts/Core/Geometry/CircleObject.d.ts
@@ -1,0 +1,25 @@
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type PositionObject from '../Renderer/PositionObject';
+
+/* *
+ *
+ *  Declarations
+ *
+ * */
+
+export interface CircleObject extends PositionObject {
+    r: number;
+}
+
+/* *
+ *
+ *  Default Export
+ *
+ * */
+
+export default CircleObject;
