+++ ts/Core/Geometry/GeometryObject.d.ts
@@ -1,0 +1,27 @@
+/* *
+ *
+ *  Imports
+ *
+ * */
+
+import type PositionObject from '../Renderer/PositionObject';
+
+/* *
+ *
+ *  Declarations
+ *
+ * */
+
+export interface GeometryObject extends PositionObject {
+    angle?: number;
+    r?: number;
+    indexes?: [number, number];
+}
+
+/* *
+ *
+ *  Default Export
+ *
+ * */
+
+export default GeometryObject;
