+++ ts/Series/Venn/VennSeries.ts
@@ -22,7 +22,10 @@
  *  Imports
  *
  * */
+
+import type CircleObject from '../../Core/Geometry/CircleObject';
 import type DataLabelOptions from '../../Core/Series/DataLabelOptions';
+import type IntersectionObject from '../../Core/Geometry/IntersectionObject';
 import type PositionObject from '../../Core/Renderer/PositionObject';
 import type ScatterPoint from '../Scatter/ScatterPoint';
 import type { SeriesStatesOptions } from '../../Core/Series/SeriesOptions';
@@ -32,28 +35,22 @@ import type SVGElement from '../../Core/Renderer/SVG/SVGElement';
 import type SVGPath from '../../Core/Renderer/SVG/SVGPath';
 import type VennPointOptions from './VennPointOptions';
 import type VennSeriesOptions from './VennSeriesOptions';
+
 import A from '../../Core/Animation/AnimationUtilities.js';
 const { animObject } = A;
 import Color from '../../Core/Color/Color.js';
 const { parse: color } = Color;
 import DrawPointMixin from '../../Mixins/DrawPoint.js';
-import GeometryMixin from '../../Mixins/Geometry.js';
+import CU from '../../Core/Geometry/CircleUtilities.js';
 const {
-    getCenterOfPoints,
-    getDistanceBetweenPoints
-} = GeometryMixin;
-import GeometryCirclesModule from '../../Mixins/GeometryCircles.js';
-const {
-    getAreaOfCircle,
     getAreaOfIntersectionBetweenCircles,
-    getCircleCircleIntersection,
     getCirclesIntersectionPolygon,
-    getOverlapBetweenCircles: getOverlapBetweenCirclesByDistance,
     isCircle1CompletelyOverlappingCircle2,
     isPointInsideAllCircles,
-    isPointInsideCircle,
     isPointOutsideAllCircles
-} = GeometryCirclesModule;
+} = CU;
+import GU from '../../Core/Geometry/GeometryUtilities.js';
+const { getCenterOfPoints } = GU;
 import NelderMeadMixin from '../../Mixins/NelderMead.js';
 const { nelderMead } = NelderMeadMixin;
 import palette from '../../Core/Color/Palette.js';
@@ -113,8 +110,8 @@ declare global {
             value: number;
         }
         interface VennUtilsObject {
-            geometry: GeometryMixin;
-            geometryCircles: object;
+            geometry: typeof GU;
+            geometryCircles: typeof CU;
             nelderMead: NelderMeadMixin;
             addOverlapToSets(
                 relations: Array<VennRelationObject>
@@ -265,14 +262,11 @@ class VennSeries extends ScatterSeries {
      * Returns the found position.
      */
     public static getLabelPosition(
-        internal: Array<Highcharts.CircleObject>,
-        external: Array<Highcharts.CircleObject>
+        internal: Array<CircleObject>,
+        external: Array<CircleObject>
     ): PositionObject {
         // Get the best label position within the internal circles.
-        let best = internal.reduce(function (
-            best: Highcharts.VennLabelPositionObject,
-            circle: Highcharts.CircleObject
-        ): Highcharts.VennLabelPositionObject {
+        let best = internal.reduce((best, circle): Highcharts.VennLabelPositionObject => {
             const d = circle.r / 2;
 
             // Give a set of points with the circle to evaluate as the best
@@ -286,10 +280,7 @@ class VennSeries extends ScatterSeries {
             ]
                 // Iterate the given points and return the one with the largest
                 // margin.
-                .reduce(function (
-                    best: Highcharts.VennLabelPositionObject,
-                    point: PositionObject
-                ): Highcharts.VennLabelPositionObject {
+                .reduce((best, point): Highcharts.VennLabelPositionObject => {
                     const margin = VennUtils.getMarginFromCircles(point, internal, external);
 
                     // If the margin better than the current best, then update
@@ -360,22 +351,21 @@ class VennSeries extends ScatterSeries {
     ): Highcharts.VennLabelValuesObject {
         const sets = relation.sets;
         // Create a list of internal and external circles.
-        const data = setRelations.reduce(function (
-            data: Record<string, (Array<Highcharts.CircleObject>)>,
-            set: Highcharts.VennRelationObject
-        ): Record<string, Array<Highcharts.CircleObject>> {
-            // If the set exists in this relation, then it is internal,
-            // otherwise it will be external.
-            const isInternal = sets.indexOf(set.sets[0]) > -1;
-            const property = isInternal ? 'internal' : 'external';
+        const data = setRelations.reduce(
+            (data, set): Record<string, Array<CircleObject>> => {
+                // If the set exists in this relation, then it is internal,
+                // otherwise it will be external.
+                const isInternal = sets.indexOf(set.sets[0]) > -1;
+                const property = isInternal ? 'internal' : 'external';
 
-            // Add the circle to the list.
-            data[property].push(set.circle);
-            return data;
-        }, {
-            internal: [],
-            external: []
-        });
+                // Add the circle to the list.
+                data[property].push(set.circle);
+                return data;
+            }, {
+                internal: [],
+                external: []
+            } as Record<string, Array<CircleObject>>
+        );
 
         // Filter out external circles that are completely overlapping all
         // internal
@@ -413,19 +403,11 @@ class VennSeries extends ScatterSeries {
     public static layout(
         relations: Array<Highcharts.VennRelationObject>
     ): ({
-            mapOfIdToShape: Record<string, (
-                Highcharts.CircleObject|Highcharts.GeometryIntersectionObject
-            )>;
-            mapOfIdToLabelValues: Record<string, (
-                Highcharts.VennLabelValuesObject
-            )>;
+            mapOfIdToShape: Record<string, (CircleObject|IntersectionObject)>;
+            mapOfIdToLabelValues: Record<string, (Highcharts.VennLabelValuesObject)>;
         }) {
-        const mapOfIdToShape: Record<string, (
-            Highcharts.CircleObject|Highcharts.GeometryIntersectionObject
-        )> = {};
-        const mapOfIdToLabelValues: Record<string, (
-            Highcharts.VennLabelValuesObject
-        )> = {};
+        const mapOfIdToShape: Record<string, (CircleObject|IntersectionObject)> = {};
+        const mapOfIdToLabelValues: Record<string, (Highcharts.VennLabelValuesObject)> = {};
 
         // Calculate best initial positions by using greedy layout.
         if (relations.length > 0) {
@@ -441,7 +423,7 @@ class VennSeries extends ScatterSeries {
                     const shape = VennUtils.isSet(relation) ?
                         mapOfIdToCircles[id] :
                         getAreaOfIntersectionBetweenCircles(
-                            sets.map((set): Highcharts.CircleObject =>
+                            sets.map((set): CircleObject =>
                                 mapOfIdToCircles[set])
                         );
 
@@ -507,7 +489,7 @@ class VennSeries extends ScatterSeries {
      */
     public static updateFieldBoundaries(
         field: Highcharts.PolygonBoxObject,
-        circle: Highcharts.CircleObject
+        circle: CircleObject
     ): Highcharts.PolygonBoxObject {
         const left = circle.x - circle.r,
             right = circle.x + circle.r,
